+++ pppine/django.py
@@ -1,5 +1,6 @@
 from django.urls import reverse
 from django.http import Http404
+from django.core.exceptions import FieldDoesNotExist
 
 
 def get_field_names(model):